Feat/optimised file writing (#15)
* Vault token renewal

Adds a new parameter for the Vault docs to drive the auth token renewal.

* Add file headers

* Complete the Vault ticket renewal.

* Avoid file to be created over and over again

For the AES decoder there was a strategy to create a file (using random UUID) every time the "key" was returned. The new code is changing the approach to make sure we only create the file once. This means though for the AES encoding the encoding format should contain an id for a file:${encoding}_${id}. The id has to be unique as well.

Small code improvement around decoding

* Fix compilation