Vault - Support default TTLs
Vault's KV2 engine does not return the `lease_duration` TTL value
that is provided by the original KV1 engine. When this value is 0
or missing it effectively disables the caching support of the
`VaultSecretProvider`.

This commit adds support for configuring an optional default TTL to apply
when the `lease_duration` value is 0.

To maintain backwards compatibility the default value remains 0.

Fixes https://github.com/lensesio/secret-provider/issues/4

Ref: https://github.com/hashicorp/vault/issues/6274