Implementing task to write maven descriptors (#66)
Java jars usually come packaged with a META-INF/maven directory, known as the maven descriptor, which contains information about all dependencies packaged within. As we were using SBT we were not generating a maven descriptor, and only passing along the descriptors from some of our dependent jars. This was causing an issue that the descriptor reported some of our components packaged within as out-of-date.

This PR adds the creation of a maven descriptor director during the compile stage, which is packaged into the final jar via sbt assembly. We also avoid packaging the maven descriptors from our dependencies as these are unreliable.

I have verified that this works by looking at the dependency report and I can see that json-smart is no longer reported as an old version.